domain #2 - 스키마 확정
1. MongoDB 스키마 확정에 따라 엔티티 설계
2. 1.1에서 발생한 이슈 해결 - 자세한 내용은 pull request 내용이나 슬랙에 표기
3. keyword에 update 추가.
